+++ CHANGELOG.md
@@ -505,3 +505,5 @@
 * Difficulty multipliers for player movement, attack, etc is now handled in RpgValues calculations
 * RpgDefaultValues can be overridden in its entirety by providing an IRpgDefaultValues implementation with RpgDefaultValuesOverrideAttribute (this is used for legacy back compat where we need the old calculations)
 * Added new placeholder texture/sprite (NULLA0) and material (DebugPattern)
+* Greatly simplified RpgDefaultValues, mostly fixed with no skill or stat definitions
+* Removed most stat and skill definitions from builtin StatType and SkillType enums